% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_settings.R
\name{add_settings}
\alias{add_settings}
\title{Add analysis settings to the blueprint}
\usage{
add_settings(data, ...)
}
\arguments{
\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{...}{
    For \code{glm}: arguments to be used to form the default
    \code{control} argument if it is not supplied directly.

    For \code{weights}: further arguments passed to or from other methods.
  }

\item{blueprint}{The blueprint object}

\item{hclust.order}{Whether to order the correlation data based on the
\code{\link[stats]{hclust}} algorithm.}

\item{cluster.id}{Variable that represents the cluster for GEE.}
}
\value{
Settings for the analysis are added to the blueprint
}
\description{
Most statistical techniques need to specify some settings for them to run.
This function sets those settings in the blueprint, before the statistical
method is used at the construction phase.
}
\examples{

design(iris, 'gee') \%>\%
 add_settings('Species', family = binomial('logit'), conf.int = FALSE)

ds <- design(iris, 'cor')
ds <- add_settings(ds, method = 'spearman')

ds <- design(iris, 't.test')
add_settings(ds, paired = TRUE)
add_settings(ds)

}

