% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variables.R
\name{add_variables}
\alias{add_variables}
\title{Add variables to the analysis}
\usage{
add_variables(blueprint, type = c("yvars", "xvars", "covariates",
  "interaction"), variables)
}
\arguments{
\item{blueprint}{The blueprint object}

\item{type}{The variable type, i.e. where it is located on the equation (y
position, x, as a covariate, etc.)}

\item{variables}{Variables to use for the type specified}
}
\value{
Adds variables to the blueprint
}
\description{
While different analyses use different types of variables, in general they
can be classified as in the 'y' or 'x' position of a statistical equation.
They can further be classified as covariates and as an interaction term.
}
\examples{

ds <- design(iris, 'cor')
ds <- add_variables(ds, 'xvar', 'Sepal.Length')
ds <- add_variables(ds, 'yvar', 'Petal.Length')

}

