% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build a mason structure}
\usage{
build(data, ...)
}
\arguments{
\item{data}{The data from the mason chain (last chain is
\code{\link{lay_base}}.}

\item{...}{Additional parameters appropriate for the \code{\link{design}}}
}
\value{
Outputs a list structure with the data, results, and additional
  design specifications for the analysis.
}
\description{
Build a mason structure
}
\examples{
ds <- data.frame(state.region, state.x77)
## GEE
design(ds, 'gee', family = gaussian, corstr = 'exchangeable') \%>\%
    lay_base(id = 'state.region', c('Income', 'Frost'), c('Population', 'Murder'), 'Life.Exp') \%>\%
    build()
## Correlation
design(ds, 'cor') \%>\%
    lay_base(c('Income', 'Frost'), c('Population', 'Murder')) \%>\%
    build()
}

