% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{Construct the results of the analysis}
\usage{
construct(blueprint, na.rm = TRUE, hclust.order = FALSE)
}
\arguments{
\item{blueprint}{The blueprint object}

\item{na.rm}{Whether to remove missing values}

\item{hclust.order}{Whether to order the correlation data based on the
\code{\link[stats]{hclust}} algorithm.}
}
\value{
Uses the blueprint to construct the results of the statistical
  analysis.
}
\description{
Construct the results of the analysis
}
\examples{

library(magrittr)
design(iris, 'cor') \%>\%
 add_settings() \%>\%
 add_variables('xvars', c('Sepal.Length', 'Sepal.Width')) \%>\%
 construct()

design(iris, 't.test') \%>\%
 add_settings() \%>\%
 add_variables('yvars', c('Sepal.Length', 'Sepal.Width')) \%>\%
 add_variables('xvars', c('Petal.Length', 'Petal.Width')) \%>\%
 construct()

}

