% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_analysis.R
\name{construct_analysis}
\alias{construct_analysis}
\title{Construct the results of the analysis}
\usage{
construct_analysis(blueprint, na.rm = TRUE, hclust.order = FALSE)
}
\arguments{
\item{blueprint}{The blueprint object}

\item{na.rm}{Whether to remove missing values}

\item{hclust.order}{Whether to order the correlation data based on the
\code{\link[stats]{hclust}} algorithm.}
}
\value{
Uses the blueprint to construct the results of the statistical
  analysis.
}
\description{
Construct the results of the analysis
}
\examples{

library(magrittr)
design_analysis(iris, 'cor') \%>\%
 add_settings() \%>\%
 add_variables('xvars', c('Sepal.Length', 'Sepal.Width')) \%>\%
 construct_analysis()

}

