% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{Design the masonry project}
\usage{
design(data, test.type, ...)
}
\arguments{
\item{data}{Raw data to process for analysis.}

\item{test.type}{The test or analysis to run.}

\item{...}{Additional parameters.}

\item{method}{Which correlation method to use.}

\item{use}{How to handle missing values in correlation.}

\item{family}{Distribution family to use for \code{\link[geepack]{geeglm}}.}

\item{corstr}{Correlation structure to use for \code{\link[geepack]{geeglm}}.}
}
\value{
Outputs a list with the raw data and specifications for use later in
  the masonry chain.
}
\description{
This is the blueprint, or design, stage of the analysis. This function sets
the data and specifications for use later in the masonry chain
(\code{\link{lay_base}}, \code{\link{build}}, and than \code{\link{polish}}).
}
\examples{
ds <- data.frame(state.region, state.x77)
## GEE
design(ds, 'gee', family = gaussian, corstr = 'exchangeable')
## Correlation
design(ds, 'cor', method = 'spearman')
}

