% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mason.R
\name{mason}
\alias{mason}
\title{Run a generator}
\usage{
mason(generator = NULL)
}
\arguments{
\item{generator}{The name of the generator to run. This is the initial
part of the package name that defines the generator.}
}
\description{
If no generator is given, then you can choose one interactively.
It will generate the application in the current working directory,
and it expects it to be empty.
}

