% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{old_polish}
\alias{old_polish}
\alias{polish}
\title{Polish the built structure}
\usage{
polish(data, ...)
}
\arguments{
\item{data}{The data object output from the mason chain (after
\code{\link{build}}).}

\item{...}{Additional arguments}
}
\value{
Outputs a single dataframe
}
\description{
Put the finishing touches to the finished structure. Used at the end of a
'mason' project (ie. at the end of a mason pipe chain). Adjust p-values for
multiple comparisons, filter the results you want, and transform the beta
coefficients if needed.
}

