% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{polish_filter}
\alias{polish_filter}
\title{Filter/subset a column by a pattern.}
\usage{
polish_filter(data, keep.pattern, column)
}
\arguments{
\item{data}{The scrubbed results.}

\item{keep.pattern}{Rows to keep based on a regular expression pattern.}

\item{column}{The column to apply the filtering to.}
}
\description{
Basically a wrapper around \code{\link[dplyr]{filter}}, but using
\code{\link[base]{grepl}} for the pattern searching.
}
\examples{
\dontrun{
ds <- scrub(ds)
polish_filter(ds, 'Xterms', 'terms')
}
}

