% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{polish_transform_estimates}
\alias{polish_transform_estimates}
\title{Transform the resulting regression estimates to something more interpretable.}
\usage{
polish_transform_estimates(data, transform.fun)
}
\arguments{
\item{data}{The scrubbed results.}

\item{transform.fun}{A function to modify continuous variable columns.}
}
\description{
Simply a thin wrapper around \code{\link[dplyr]{mutate_each}}.
}
\examples{
\dontrun{
ds <- scrub(ds)
polish_transform_estimates(ds, function(x) exp(x))
}
}

