% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub}
\alias{scrub}
\title{Scrub down and polish up the constructed analysis results.}
\usage{
scrub(blueprint)
}
\arguments{
\item{blueprint}{The blueprint object.}
}
\value{
Outputs a cleaned up version of the constructed analysis.
}
\description{
Scrub down and polish up the constructed analysis results.
}
\examples{

ds <- design(iris, 'cor')
ds <- add_settings(ds)
ds <- add_variables(ds, 'xvars', c('Sepal.Length', 'Sepal.Width'))
ds <- construct(ds)
scrub(ds)

}

