% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_settings.R
\name{add_settings}
\alias{add_settings}
\title{Add analysis settings to the blueprint}
\usage{
add_settings(blueprint, family, corstr, cluster.id, conf.int, conf.level,
  method, use)
}
\arguments{
\item{blueprint}{The blueprint object}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link{family}} for details of
    family functions.)}

\item{cluster.id}{Variable that represents the cluster for GEE}

\item{conf.int}{whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if
\code{conf.int=TRUE}}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\value{
Settings for the analysis are added to the blueprint
}
\description{
Most statistical techniques need to specify some settings for them to run.
This function sets those settings in the blueprint, before the statistical
method is used at the construction phase.
}
\examples{

ds <- design_analysis(iris, 'gee')
ds <- add_settings(family = binomial('logit'), conf.int = FALSE)

ds <- design_analysis(iris, 'cor')
ds <- add_settings(method = 'spearman')

}

