% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\alias{build.cor_df}
\alias{build.gee_df}
\title{Build a mason structure}
\usage{
build(data, ...)

\method{build}{gee_df}(data, conf.int = TRUE, conf.level = 0.95)

\method{build}{cor_df}(data)
}
\arguments{
\item{data}{The data from the mason chain (previous chain is
\code{\link{lay_base}}.}

\item{...}{Additional parameters appropriate for the \code{\link{design}}}

\item{conf.int}{Whether to include the confidence interval}

\item{conf.level}{Range for the confidence interval}
}
\value{
Outputs a list structure with the data, results, and additional
  design specifications for the analysis.
}
\description{
Build a mason structure
}
\examples{
ds <- data.frame(state.region, state.x77)
## GEE
design(ds, 'gee', family = gaussian, corstr = 'exchangeable') \%>\%
    lay_base(id = 'state.region', c('Income', 'Frost'), c('Population', 'Murder'), 'Life.Exp') \%>\%
    build()
## Correlation
design(ds, 'cor') \%>\%
    lay_base(c('Income', 'Frost'), c('Population', 'Murder')) \%>\%
    build()
ds <- data.frame(state.region, state.x77)
## Correlation
design(ds, 'cor') \%>\%
    lay_base(c('Income', 'Frost'), c('Population', 'Murder')) \%>\%
    build()
}

