% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{.fetch_design.cor}
\alias{.fetch_design.gee}
\alias{.fetch_design.lm}
\alias{design}
\title{Design the masonry project}
\usage{
design(data, test.type, ...)

\method{.fetch_design}{gee}(data, test.type, family = gaussian("identity"),
  corstr = c("independence", "exchangeable", "ar1"))

\method{.fetch_design}{cor}(data, test.type, method = c("pearson", "kendall",
  "spearman"), use = c("complete.obs", "all.obs", "pairwise.complete.obs",
  "everything", "na.or.complete"))

\method{.fetch_design}{lm}(data, test.type)
}
\arguments{
\item{data}{Raw data to process for analysis.}

\item{test.type}{The test or analysis to run.}

\item{...}{Additional parameters.}

\item{family}{Distribution family to use for \code{\link[geepack]{geeglm}}.}

\item{corstr}{Correlation structure to use for \code{\link[geepack]{geeglm}}.}

\item{method}{Which correlation method to use.}

\item{use}{How to handle missing values in correlation.}
}
\value{
Outputs a list with the raw data and specifications for use later in
  the masonry chain.
}
\description{
This is the blueprint, or design, stage of the analysis. This function sets
the data and specifications for use later in the masonry chain
(\code{\link{lay_base}}, \code{\link{build}}, and than \code{\link{polish}}).
}
\examples{

ds <- data.frame(state.region, state.x77)
## GEE
design(ds, 'gee', family = gaussian, corstr = 'exchangeable')

ds <- data.frame(state.region, state.x77)
## Correlation
design(ds, 'cor', method = 'spearman')

## lm
ds <- data.frame(state.region, state.x77)
design(ds, 'lm')

}

