% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{polish}
\alias{polish}
\alias{polish_adjust_pvalue}
\alias{polish_renaming}
\title{Polish up.}
\usage{
polish_renaming(data, renaming.fun, columns = NULL)

polish_adjust_pvalue(data, method = "BH")
}
\arguments{
\item{data}{The scrubbed object.}

\item{renaming.fun}{A function, typically with \code{\link[base]{gsub}}, that
searches and replaces strings.}

\item{columns}{The columns to apply the renaming function to. Defaults to
columns that are a factor or character vectors.}

\item{method}{correction method.  Can be abbreviated.}

\item{data}{The scrubbed results.}
}
\description{
Polish up.

Polish up the scrubbed results by renaming X and Y term names.

A thin wrapper around \code{\link[dplyr]{mutate}} and \code{\link[stats]{p.adjust}}
}
\examples{
\dontrun{
ds <- scrub(ds)
polish_renaming(ds, function(x) gsub('original string', 'replacement word', x))
}
}

